AuthCheck/requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
flask>=2.3.0
flask-sqlalchemy>=3.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
requests>=2.31.0
cryptography>=41.0.0